# optimize large slice alloc in Stats