fix #104

Introduce sync pool to optimize large object allocation

```
$ benchstat bench.out.master bench.out.perf
goos: linux
goarch: amd64
pkg: github.com/safchain/ethtool
cpu: Intel(R) Core(TM) i7-1065G7 CPU @ 1.30GHz
               │ bench.out.master │            bench.out.perf            │
               │      sec/op      │    sec/op     vs base                │
Stats/enp0s3-4       789.0µ ± 11%   398.5µ ± 14%  -49.49% (p=0.002 n=10)
Stats/enp0s8-4       777.9µ ±  8%   373.7µ ±  7%  -51.96% (p=0.000 n=10)
geomean              783.4µ         385.9µ        -50.74%

               │ bench.out.master │            bench.out.perf             │
               │       B/op       │     B/op       vs base                │
Stats/enp0s3-4    1300.243Ki ± 0%   3.576Ki ± 30%  -99.72% (p=0.000 n=10)
Stats/enp0s8-4    1300.238Ki ± 0%   3.896Ki ± 11%  -99.70% (p=0.000 n=10)
geomean              1.270Mi        3.732Ki        -99.71%

               │ bench.out.master │           bench.out.perf           │
               │    allocs/op     │ allocs/op   vs base                │
Stats/enp0s3-4         59.00 ± 0%   51.00 ± 0%  -13.56% (p=0.000 n=10)
Stats/enp0s8-4         59.00 ± 0%   51.00 ± 0%  -13.56% (p=0.000 n=10)
geomean                59.00        51.00       -13.56%
```