@ritwikranjan 2025-05-07
This is the change i was planning: https://github.com/safchain/ethtool/pull/106, maybe we can combine the two of them. This change can be done as a default and then we would also expose those buffers for customers to have them manage those variables themselves.