@halfcrazy 2025-05-07
> This is the change i was planning: #106, maybe we can combine the two of them. This change can be done as a default and then we would also expose those buffers for customers to have them manage those variables themselves.

That's ok. The original `Stats` will use the in-tree buffer pool.